Initial refactoring to enable same pattern searches for dissolved companies
